+++ src/Ui/Bundle/AdminBundle/Resources/assets/js/components/_BatchSelectionHelper.js
@@ -80,4 +80,4 @@ BatchSelectionHelper.prototype.triggerChange = function (totalSelection) {
     this.batchForm.dispatchEvent(htmlEvent);
 };
 
-module.exports = BatchSelectionHelper;
+export default BatchSelectionHelper;