+++ src/Ui/Bundle/AdminBundle/Resources/assets/js/components/_PreventMultipleSubmit.js
@@ -16,4 +16,4 @@ PreventMultipleSubmit.prototype.onSubmit = function ()
     }.bind(this), 1);
 };
 
-module.exports = PreventMultipleSubmit;
+export default PreventMultipleSubmit;