+++ src/Ui/Bundle/EventBundle/Resources/assets/js/components/VideoConference/Subscriber.js
@@ -97,4 +97,4 @@ Subscriber.prototype.handleError = function (error) {
   }
 };
 
-module.exports = Subscriber;
+export default Subscriber;