+++ CHANGELOG.md
@@ -4,6 +4,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-2217 - Activate / deactivate the networking tab by key date
+- VIMEET-2204 - Hidden contact menu
+
+## [2.47.0] - 2020-09-11
+### Added 
+- VIMEET-2226 - Export multi-upload and media with sheet, add asynchronous export
+- VIMEET-2156 - Remove explicit error about email on login (security)
 
 ## [2.46.0] - 2020-09-10
 ### Added